Regularize *_language to *_lang_id
XML attributes of the form `lang_id=...` have sometimes been mapped to `..._lang_id`, and sometimes to `..._language`.

Since in one case (`grant_agency_lang_id`) there was a mismatch between the value in `wos_config.xml` (which had `...lang_id`) and that in `wos_schema.sql` (which had `...language`), an edit was necessary, so I thought I'd take the opportunity to regularize them all.